Implementacion modulo proveedores, registro de compras y diseno glassmorphic de login con logo Z animado

+++ sistemaVentas/src/main/java/com/sistemaVentas/dataaccess/CategoriaRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.sistemaVentas.entity.Categoria;
 
 @Repository
-public interface CategoriaRepository extends JpaRepository<Categoria, Long> {
+public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
 }
